requirements.txt
pillow==10.4.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.26.0
GitPython==3.1.43
gitdb==4.0.11

# JSON handling and validation
jsonschema==4.23.0
jsonschema-specifications==2023.12.1

# Job management and tools
joblib==1.4.2
filelock==3.16.1

# Web utilities and other core libraries
requests==2.32.3
Jinja2==3.1.4
MarkupSafe==2.1.5
click==8.1.7
blinker==1.8.2
attrs==24.2.0

# Miscellaneous tools
altair==5.4.1
scikit-learn==1.5.2
scipy==1.13.1

# Huggingface transformers and PyTorch for LLM usage
transformers==4.33.3
torch==2.0.1

# Additional core dependency
setuptools>=60.0.0  # Fixes distutils
echo "sentence-transformers" >> requirements.txt
